python/stock_get/requirements.txt
xlrd
pandas_datareader
GitPython